[wip] organize and document things for code review
