+++ src/zql/index.ts
@@ -1,0 +1,2 @@
+export {EntityQuery} from './query/EntityQuery.js';
+export {Context} from './context/context.js';
