+++ src/zql/ivm/source/ISource.ts
@@ -12,6 +12,7 @@ export interface ISourceInternal {
   onCommitPhase1(version: Version): void;
   // Drain queues and run the reactive graph
   // TODO: we currently can't rollback a transaction in phase 2
+  // as in, if an operator fails in the graph we're in a partial state that can't be reverted.
   onCommitPhase2(version: Version): void;
   // Now that the graph has computed itself fully, notify effects / listeners
   onCommitted(version: Version): void;
